fix: changed argument type to uint256

+++ src/ReservoirPair.sol
@@ -56,7 +56,7 @@ abstract contract ReservoirPair is AssetManagedPair, OracleWriter, ReentrancyGua
         emit Sync(reserve0, reserve1);
     }
 
-    function setMaxChangePerSecond(uint64 aChangePerSecond) external override onlyFactory {
+    function setMaxChangePerSecond(uint256 aChangePerSecond) external override onlyFactory {
         require(0 < aChangePerSecond && aChangePerSecond <= MAX_CHANGE_PER_SEC, "RP: INVALID_CHANGE_PER_SECOND");
         allowedChangePerSecond = aChangePerSecond;
     }
